requirements.txt
numpy==1.21.5
opencv-python==4.5.5.64
face_recognition==1.3.0
Pillow==9.0.1
tk==0.1.0
setuptools==60.5.0
dlib==19.23.1
imutils==0.5.4
